+++ ecg_classifier/configs/criterion/cr.yaml
@@ -1,4 +1,3 @@
 # cr.yaml
 
 _target_: torch.nn.BCEWithLogitsLoss
-pos_weight: ???
