+++ ecg_classifier/configs/early_stopping/es.yaml
@@ -1,4 +1,4 @@
 # es.yaml
 
-_target_: ecg_classifier.EarlyStopping
+_target_: ecg_classifier.module.EarlyStopping
 patience: 10
