+++ ecg_classifier/main.py
@@ -3,26 +3,22 @@ import hydra
 from hydra.utils import instantiate
 from omegaconf import DictConfig
 
-from ecg_classifier import train, test, device
+from ecg_classifier.module import train, test
+from ecg_classifier.utils import device
 
 
 @hydra.main(config_path="configs/", config_name="config.yaml", version_base=hydra.__version__)
 def main(cfg: DictConfig):
-    target_labels = cfg.dataset.target_labels
-
     ecg_dataset = instantiate(cfg.dataset)
 
     train_dataset, val_dataset, test_dataset = ecg_dataset.get_dataset()
-    cfg.criterion.pos_weight = ecg_dataset.get_pos_weight().to(device)
+    pos_weight = ecg_dataset.get_pos_weight().to(device)
     ecg_dataset.close_dataset()
 
     batch_size  = cfg.model.train.batch_size
     n_epoch     = cfg.model.train.n_epoch
-    num_classes = cfg.model.resnet.num_classes
 
-    threshold_preds = [0.5] * num_classes
-
-    criterion = instantiate(cfg.criterion)
+    criterion = instantiate(cfg.criterion, pos_weight=pos_weight)
 
     train_loader = torch.utils.data.DataLoader(train_dataset, batch_size=batch_size)
     val_loader   = torch.utils.data.DataLoader(val_dataset,   batch_size=batch_size)
@@ -35,7 +31,7 @@ def main(cfg: DictConfig):
 
     net, threshold_preds, train_history, val_history = train(net, train_loader, val_loader, 
                                                              n_epoch, optimizer, criterion, 
-                                                             scheduler, early_stopping, threshold_preds)
+                                                             scheduler, early_stopping)
     test_loss = test(net, test_loader, criterion, threshold_preds)
