+++ ecg_classifier/utils/device.py
@@ -1,0 +1,4 @@
+import torch
+
+device = 'cuda' if torch.cuda.is_available() else 'cpu'
+print(f"Training will take on {device}")
